@seemethere 2017-02-12
NOTE: This does not bring us back to previous speeds but it does bring a speed boost we still need to figure out ways to eek out more speed without squashing default headers that the user wants to set.